requirements.txt
py-cord==2.3.2
python-dotenv==0.15.0
youtube-dl==2021.12.17
PyNaCl==1.5.0
nest-asyncio==1.5.6
pymongo==4.3.2